+++ app/src/main/java/dk/schulz/voiceme/ui/theme/Type.kt
@@ -1,0 +1,5 @@
+package dk.schulz.voiceme.ui.theme
+
+import androidx.compose.material3.Typography
+
+val Typography = Typography()